Refactor models in server module
